Repaired some things in response to Frank Fertitta's email on sept. 2005:
Frank,

> > what your intention is here. It appears as though you are
> > just energizing the windings after you stop. If this is so,
> > what's keeping the motor from advancing to the next position
> > to reach "1111"? If the last state sent to the motor was

Yes, this is not the best thing to do.  I will update cvs this evening, hopefully.

What it should be doing is to simply leave the outputs in the last state unless provide_static_holding is '0', in which case it should turn all the outputs off.

outputting "1111" will not cause the motor to move, because it will cause the two sets of windings to "arm wrestle" each other, but it's not the best idea because it wastes power aka. heat.

Best,
Steve

> Hello Steve,
>
> I was looking at your code you wrote for controlling a 4 or 6
> wire stepper motor. I see how you have added what could be a
> means for providing holding current but I don't understand
> what your intention is here. It appears as though you are
> just energizing the windings after you stop. If this is so,
> what's keeping the motor from advancing to the next position
> to reach "1111"? If the last state sent to the motor was
> "1010" then you stop at "1111" the motor is going to move. Or
> is your code only going to stop at "1111"? If that's the
> case, then will this limit the resolution of the motor? As I
> see it your code doesn't stop at "1111" unless it just
> happens to stop there when the count is "0".
> Am I missing something?
>
> Thanks for your excellent contributions. I do indeed learn
> from reading your work.
>
> Frank Fertitta